+++ subcommand/mesh-init/command_test.go
@@ -67,12 +67,14 @@ func TestRun(t *testing.T) {
 		expAdditionalMeta map[string]string
 		serviceName       string
 		expServiceName    string
+		proxyPort         int
 
 		consulLogin config.ConsulLogin
 	}{
 		"basic service": {},
 		"service with port": {
 			servicePort: 8080,
+			proxyPort:   21000,
 		},
 		"service with upstreams": {
 			upstreams: []config.Upstream{
@@ -153,6 +155,7 @@ func TestRun(t *testing.T) {
 	}
 
 	for name, c := range cases {
+		c := c
 		t.Run(name, func(t *testing.T) {
 			var (
 				taskARN          = "arn:aws:ecs:us-east-1:123456789:task/test/abcdef"
@@ -182,6 +185,10 @@ func TestRun(t *testing.T) {
 
 			if c.expServiceName != "" {
 				expectedServiceName = c.expServiceName
+			}
+
+			if c.proxyPort == 0 {
+				c.proxyPort = config.DefaultPublicListenerPort
 			}
 
 			for i := range c.upstreams {
@@ -233,7 +240,8 @@ func TestRun(t *testing.T) {
 				HealthSyncContainers: nil,
 				ConsulLogin:          c.consulLogin,
 				Proxy: &config.AgentServiceConnectProxyConfig{
-					Upstreams: c.upstreams,
+					PublicListenerPort: c.proxyPort,
+					Upstreams:          c.upstreams,
 				},
 				Service: config.ServiceRegistration{
 					Name:   c.serviceName,
@@ -269,7 +277,7 @@ func TestRun(t *testing.T) {
 			expectedProxyServiceRegistration := &api.AgentService{
 				ID:      expSidecarServiceID,
 				Service: fmt.Sprintf("%s-sidecar-proxy", expectedServiceName),
-				Port:    20000,
+				Port:    c.proxyPort,
 				Kind:    api.ServiceKindConnectProxy,
 				Proxy: &api.AgentServiceConnectProxyConfig{
 					DestinationServiceName: expectedServiceName,